+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 0.4.0 (2026-08-07)
 
 - Skip runtime refreshes for changes limited to ordinary PHP method bodies